+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 
-## 1.0.0-beta.1 - 2019-02-10
+## 1.0.2 - 2019-06-17
+
+### Fixed
+- Fixed invalid message category ([#11])
+
+[#11]: https://github.com/barrelstrength/craft-sprout-redirects/issues/11
+
+## 1.0.1 - 2019-04-28
+
+### Changed
+- Updated barrelstrength/sprout-base-redirects requirement v1.0.10
+
+### Fixed
+- Improved Postgres support
+
+## 1.0.0 - 2019-04-24
 
 ### Added 
-- Added initial release
+- Initial release